Moving lib into subdir (you can't have a package called go...)
