+++ go/lib/kitten_test.go
@@ -1,4 +1,4 @@
-package kittenlib
+package lib
 
 import (
 	"testing"
